Cast the return value to a float for the getDouble method.
Indeed, for really small or really large numbers, the mysql database return the result as a string like this : -1.98526e+16 and getDouble is suppose to return a float value and not a string.

This fix an issue with the Zend_Measure constructor which fail with string like this : "-1.98526e+16" in parameter as oppose to the float value.